Simplify GeoJSONWorkerSource
